Update script.js
Press S to download all videos in the course starting from the course which is currently playing.
Press G to download a single video which is currently playing.

*Notes:
In auto download mode, videos will download after a gap of 50 seconds. This is done to avoid PluralSight detection of high usage.